refactor(console): prompts controller (wave 3)
Extracts the Console prompt cluster out of `ChatScreen` into
`UI/Console_Modules/prompts.py` (`ConsolePromptsController`): the
source-aware Prompt Library modal (`_open_console_prompts_modal`, 397
lines -- the largest non-`__init__` method the class had), `/prompt` and
`/system` name resolution and their shared refuse-a-Recipe guard, both
prompt-picker launches, the system-prompt editor and its save-to-Library
flow, the Library "Use in Console" staged-insert handoff, and the shared
prompt-history store. 13 methods move; 6 keep a 3-line delegation on the
screen (outside callers: the composer menu, the command-registry dict,
`action_open_console_system_prompt_editor`, two `set_timer` schedules,
and tests that replace those exact names), 7 leave no residue.

Binding follows waves 1-2 (`dictation.py`'s canonical `__init__`
docstring, `message.py`'s shape): `run_worker`/`push_screen` live-read via
`@property`; 18 named keyword-only dependencies, each wired at the
construction site as a late-binding lambda -- eleven are replaced by name
on the screen instance somewhere in the pre-existing suite, so a
constructor snapshot would silently stop observing them. Three of the
eighteen are the session seam (`_ensure_active_console_session_settings`,
`_apply_console_session_system_prompt`, `_sync_console_session_draft`),
resolved at call time so controller construction order never matters.

Zero DOM in the controller: `_insert_prompt_text_into_composer` does a raw
`query_one`, so it stays on the screen and is injected instead. Four other
`*prompt*`-named methods stay too, each a name-pattern false positive
(RAG evidence counting, the canned caption insert, the `/rewind` row
builder) or a Textual-dispatch requirement (`action_*`, `@on`).

`chat_screen.py`: 19,662 -> 18,903 lines (-759), 605 -> 598 methods.
pyflakes output on the screen is byte-identical to baseline (18 imports
orphaned by the move were removed; nothing new appeared) and the new
module is clean.

Three test files needed real edits rather than incidental ones:
- `test_console_composer_history.py` patched `chat_screen.default_prompt_
  history_path`; the reader moved, so the patch target moved with it or it
  would have silently written to the user's real history path.
- `test_console_command_composer.py` called three of the moved methods
  UNBOUND with hand-built `SimpleNamespace` selves; retargeted to the
  controller (the `/system` one drops its `_session` namespace for the
  controller's own same-named seam property).
- `test_application_state_ownership.py` AST-checks `_consume_pending_
  console_prompt_insert`'s log statements by class+name; repointed at the
  controller so it keeps checking the real body rather than a delegation.
