+++ Tests/UI/test_console_composer_history.py
@@ -326,7 +326,7 @@ async def test_console_ghost_text_renders_and_right_arrow_accepts(
     history_path = tmp_path / "prompt_history.jsonl"
     _seed_history_file(history_path, "explain quantum computing")
     monkeypatch.setattr(
-        "tldw_chatbook.UI.Screens.chat_screen.default_prompt_history_path",
+        "tldw_chatbook.UI.Console_Modules.prompts.default_prompt_history_path",
         lambda: history_path,
     )
     app = _build_test_app()
@@ -373,7 +373,7 @@ async def test_console_up_down_recall_gated_to_boundary_rows(tmp_path, monkeypat
     history_path = tmp_path / "prompt_history.jsonl"
     _seed_history_file(history_path, "first prompt", "second prompt")
     monkeypatch.setattr(
-        "tldw_chatbook.UI.Screens.chat_screen.default_prompt_history_path",
+        "tldw_chatbook.UI.Console_Modules.prompts.default_prompt_history_path",
         lambda: history_path,
     )
     app = _build_test_app()
@@ -429,7 +429,7 @@ async def test_console_send_records_to_shared_prompt_history(tmp_path, monkeypat
 
     history_path = tmp_path / "prompt_history.jsonl"
     monkeypatch.setattr(
-        "tldw_chatbook.UI.Screens.chat_screen.default_prompt_history_path",
+        "tldw_chatbook.UI.Console_Modules.prompts.default_prompt_history_path",
         lambda: history_path,
     )
     gateway = CapturingGateway()
